something approaching a full nav graph

+++ src/simulation/world/chunk/node.rs
@@ -1,8 +1,9 @@
 use crate::simulation::world::chunk;
 use glam::IVec3;
 
+#[derive(Clone, Debug)]
 pub struct Node {
     pub grid_position: IVec3,
-    pub clearance: usize,
+    pub clearance: i32,
     pub edge_list: Vec<chunk::Edge>,
 }
